Fixed: sphere crashed at the client login on Linux if the build was compiled with O2 or O3 optimization flags. This was caused by an *extremely* messy TwoFish implementation, which i changed with a more simple (and working) one. Curious that it doesn't crash without optimization flags... Refactored CEncrypt: renamed the class to CCrypto, divided its implementationinto multiple cpp files, little cleanup and some comments added. So needed. Fixed a bug introduced with the ACTARG commit (two commits ago), which caused magery and .tele command to not work well (the spell number was cleared before starting to launch the spell). Fixed/suppressed some more warnings.

+++ src/network/send.cpp
@@ -3079,10 +3079,10 @@ void PacketServerList::writeServerEntry(const CServerRef& server, int index, boo
 
 	uint percentFull;
 	if (server == &g_Serv)
-		percentFull = minimum((server->StatGet(SERV_STAT_CLIENTS) * 100) / maximum(1, g_Cfg.m_iClientsMax), 100);
+		percentFull = (uint)(minimum((server->StatGet(SERV_STAT_CLIENTS) * 100) / maximum(1, g_Cfg.m_iClientsMax), 100));
 		//percentFull = (int)maximum(0, minimum((server->StatGet(SERV_STAT_CLIENTS) * 100) / maximum(1, g_Cfg.m_iClientsMax), 100));
 	else
-		percentFull = minimum(server->StatGet(SERV_STAT_CLIENTS), 100);
+		percentFull = (uint)(minimum(server->StatGet(SERV_STAT_CLIENTS), 100));
 
 	dword ip = server->m_ip.GetAddrIP();
